# backend/src/modules/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MikConnect API",
    "version": "1.0.0",
    "description": "API de gestion de hotspots WiFi, tickets, revendeurs et paiements pour MikConnect.",
    "contact": { "name": "MikConnect Support", "email": "support@mikconnect.com" }
  },
  "servers": [{ "url": "/api", "description": "API v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Hotspot": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "location": { "type": "string" }, "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }, "routerIp": { "type": "string" }, "routerApiPort": { "type": "integer" } } },
      "Plan": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "price": { "type": "number" }, "duration": { "type": "integer" }, "durationUnit": { "type": "string", "enum": ["minutes", "hours", "days"] }, "bandwidth": { "type": "integer" }, "dataLimit": { "type": "integer" } } },
...
      "Reseller": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "phone": { "type": "string" }, "email": { "type": "string" }, "commission": { "type": "number" } } },
      "Partner": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "type": { "type": "string", "enum": ["TECHNOLOGY", "ISP", "HARDWARE", "OTHER"] }, "contactName": { "type": "string" }, "contactEmail": { "type": "string" }, "commission": { "type": "number" } } },
...
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Creer un compte",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name","email","password"], "properties": { "name": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string", "minLength": 6 }, "phone": { "type": "string" }, "country": { "type": "string" }, "companyName": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Compte cree" }, "400": { "description": "Email deja utilise" } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Connexion",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["email","password"], "properties": { "email": { "type": "string" }, "password": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Token JWT + profil" }, "401": { "description": "Identifiants invalides" } }
      }
    },
    "/hotspots": {
      "get": {
        "tags": ["Hotspots"],
        "summary": "Lister mes hotspots",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Liste des hotspots", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Hotspot" } } } } } }
      },
      "post": {
        "tags": ["Hotspots"],
        "summary": "Creer un hotspot",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Hotspot" } } } },
        "responses": { "201": { "description": "Hotspot cree" } }
      }
    },
    "/plans": {
      "get": {
        "tags": ["Plans"],
        "summary": "Lister les offres",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Liste des offres", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Plan" } } } } } }
      },
      "post": {
        "tags": ["Plans"],
        "summary": "Creer une offre",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Plan" } } } },
        "responses": { "201": { "description": "Offre creee" } }
      }
    },
...
    "/payments/initiate": {
      "post": {
        "tags": ["Paiements"],
        "summary": "Initier un paiement mobile money",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["amount","method","phoneNumber"], "properties": { "amount": { "type": "number" }, "method": { "type": "string", "enum": ["MTN_MOMO","MOOV","WAVE","ORANGE_MONEY"] }, "phoneNumber": { "type": "string" }, "planId": { "type": "string" }, "hotspotId": { "type": "string" }, "customerName": { "type": "string" }, "customerEmail": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Paiement initie" } }
      }
    },
    "/payments/webhook/momo": {
      "post": {
        "tags": ["Paiements"],
        "summary": "Webhook mobile money (interne)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "reference": { "type": "string" }, "status": { "type": "string", "enum": ["SUCCESS","FAILED"] } } } } } },
        "responses": { "200": { "description": "Webhook recu" } }
      }
    },
...
    "/resellers": {
      "get": {
        "tags": ["Revendeurs"],
        "summary": "Lister les revendeurs",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Liste des revendeurs", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Reseller" } } } } } }
      },
      "post": {
        "tags": ["Revendeurs"],
        "summary": "Ajouter un revendeur",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reseller" } } } },
        "responses": { "201": { "description": "Revendeur cree" } }
      }
    },
    "/partners": {
      "get": {
        "tags": ["Partenaires"],
        "summary": "Lister les partenaires",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Liste des partenaires", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Partner" } } } } } }
      },
      "post": {
        "tags": ["Partenaires"],
        "summary": "Ajouter un partenaire",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Partner" } } } },
        "responses": { "201": { "description": "Partenaire cree" } }
      }
    },
...
    "/mikrotik/test": {
      "post": {
        "tags": ["MikroTik"],
        "summary": "Tester la connexion a un routeur",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "host": { "type": "string" }, "port": { "type": "integer", "default": 8728 }, "username": { "type": "string" }, "password": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Resultat du test" } }
      }
    },
...
  },
...
}